remove stale files
